LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/week-expected.txt


PASS empty value 
PASS Valid value: Value should be 2014-W52 
FAIL 2014 has 52 weeks: Value should be empty assert_equals: expected "" but got "2014-W53"
PASS 2015 has 53 weeks: Value should be 2015-W53 
FAIL Invalid value: year only assert_equals: expected "" but got "2014"
FAIL Invalid value: no week number assert_equals: expected "" but got "2014W"
FAIL Invalid value: no '-' (U+002D) assert_equals: expected "" but got "2014W52"
FAIL Invalid value: yearless week assert_equals: expected "" but got "-W52"
FAIL Invalid value: yearless week and no '-' (U+002D) assert_equals: expected "" but got "W52"
PASS Value >= min attribute 
FAIL Value < min attribute assert_equals: expected "2014-W02" but got "2014-W01"
PASS Value <= max attribute 
FAIL Value > max attribute assert_equals: expected "2014-W11" but got "2014-W12"
